+++ plugins/processor/site.mjs
@@ -1,60 +1,28 @@
 import { ReflectionKind } from 'typedoc';
-import { fullName } from '../shared/titles.mjs';
+import { CATEGORY_RULES } from '../shared/categories.mjs';
 
-const ROOT_GROUP = 'webpack';
+const SIDEBAR_KINDS =
+  ReflectionKind.Project | ReflectionKind.Namespace | ReflectionKind.Class;
 
-const toOutputPath = url => {
-  const withoutExtension = url.replace(/\.md$/, '');
-  if (withoutExtension === 'README') return '';
-  return withoutExtension.replace(/\/index$/, '');
-};
+const SIDEBAR_GROUP_NAME = 'API Documentation';
+
+const getFirstAtxHeading = text => text.match(/^#\s+(.+)$/m)?.[1]?.trim();
+
+const getFirstPathSegment = url => url.replace(/^\//, '').split('/')[0];
 
 const toSidebarLink = url => {
-  const path = toOutputPath(url);
+  const path = url.replace(/\.md$/, '').replace(/\/index$/, '');
   return path ? `/${path}` : '/';
 };
 
-const pagePathParts = url =>
-  url.replace(/\.md$/, '').split('/').filter(Boolean);
-
-const groupKeyFor = url => {
-  const parts = pagePathParts(url);
-  return parts.length > 1 ? parts[0] : ROOT_GROUP;
-};
-
-const rawPageName = url => {
-  if (url === 'README.md') return ROOT_GROUP;
-  return (
-    pagePathParts(url)
-      .at(-1)
-      ?.replace(/\/index$/, '') ?? ROOT_GROUP
-  );
-};
-
-const stripWebpackPrefix = value => value.replace(/^webpack\.?/, '');
-
-const trimGroupPrefix = (value, groupKey) => {
-  if (value === groupKey) return rawPageName(`${groupKey}/index.md`);
-  return value.startsWith(`${groupKey}.`)
-    ? value.slice(groupKey.length + 1)
-    : value;
-};
-
-const itemLabelFor = (target, url, groupKey) => {
-  const name = stripWebpackPrefix(fullName(target));
-  const label = trimGroupPrefix(name, groupKey);
-  return label || rawPageName(url);
+const defaultLabelFor = (target, url) => {
+  if (url.endsWith('/index.md')) return 'Overview';
+  if (url.endsWith('/types.md')) return 'Types';
+  return target.name;
 };
 
 const isSidebarTarget = (router, target) => {
-  if (
-    !target.kindOf?.(
-      ReflectionKind.Project | ReflectionKind.Namespace | ReflectionKind.Class
-    )
-  ) {
-    return false;
-  }
-
+  if (!target.kindOf?.(SIDEBAR_KINDS)) return false;
   if (!router.hasOwnDocument(target)) return false;
 
   const url = router.getFullUrl(target);
@@ -62,30 +30,45 @@ const isSidebarTarget = (router, target) => {
 };
 
 export const sidebar = router => {
-  const groups = new Map();
+  const categories = new Map();
+  const seen = new Set();
 
   for (const target of router.getLinkTargets()) {
     if (!isSidebarTarget(router, target)) continue;
 
     const url = router.getFullUrl(target);
-    const groupKey = groupKeyFor(url);
-    const group = groups.get(groupKey) ?? [];
-    const link = toSidebarLink(url);
+    if (seen.has(url)) continue;
+    seen.add(url);
 
-    if (!group.some(item => item.link === link)) {
-      group.push({
-        link,
-        label: itemLabelFor(target, url, groupKey),
-      });
-    }
+    const firstPathSegment = getFirstPathSegment(url);
+    const category = CATEGORY_RULES.find(
+      cat => cat.category === firstPathSegment
+    );
 
-    groups.set(groupKey, group);
+    if (!category) continue;
+
+    const headingName = getFirstAtxHeading(
+      target.comment?.summary?.[0]?.text ?? ''
+    );
+    const label = headingName ?? defaultLabelFor(target, url);
+
+    const group = categories.get(category.category) ?? {
+      label: category.label,
+      items: [],
+    };
+
+    group.items.push({
+      link: toSidebarLink(url),
+      label,
+    });
+
+    categories.set(category.category, group);
   }
 
-  return [...groups]
-    .map(([groupKey, items]) => ({
-      groupName: groupKey,
-      items,
-    }))
-    .filter(group => group.items.length);
+  return [
+    {
+      groupName: SIDEBAR_GROUP_NAME,
+      items: [...categories.values()].filter(category => category.items.length),
+    },
+  ];
 };
